Optimize AVX2 by always selecting AffineTransform specialization for small inputs. Move simd.h from src to src/nnue/layers No functional change bench: 5609606
